Refactor SongTimeline scrolling elements into abstract class
Future commits will add a second timeline view for sequencing notes
in NoteTracks and much of the code will be similar to the code in
SongTimeline. The abstract class that has been created implements
much of the code required to do scrolling of the bars and object pooling,
and what is left (usually things that require a concrete implementation of
ScrollableBar) is relatively simple to implement.