Refactor pointer code
This is a large refactor of all pointer code in the View.
Before a top-down system was used, where the view at the top would
pass events to the timeline, which would pass events to objects.
Clearly, this is inefficient when pixi supports bottom-up event propagation.
This creates a few helper classes to support clicking and dragging,
and changes everything to use this bottom-up system.