Finish timeline view switching, implement auto resizing
This commit completes the code required to allow the timelines
to switch correctly, including mousewheelevents transitioning
between timeline views correctly. Also, the PIXI app has
been moved inside a React object, and the entire app functions
out of SequencerApp. This comes with the effect that it is completely
full screen, and will adjust to browser resize events correctly.