Add drag and drop to effects chain list, other changes
This commit mainly adds react-beautiful-dnd as a dependency,
but from here onwards react, react-dom and react-beautiful-dnd
are externals, allowing the project to benefit from cached versions
of react, etc. This reduces the webpack bundle size significantly.

This commit also fixes a bug that appeared as a result of this
change where inline changes to top, left, width, height, etc
would not be applied without appending "px" to the string.