Add saving and loading to UI
This commit adds Redux as a dependency to the project.
This was used to solve the issue of having a single point of truth
within the UITracks. By using redux, any part of the UI can add, remove
or set the UI tracks, and this will immediately be reflected
across the UI. Note the implementation of Redux is not entirely correct
as it is only used at a high level (adding and removing tracks) and
the immutability of the objects is broken for track attribute changes.
This has been done as it would require rewriting large portions
of the system, including most/all of UITrack and a lot of
the SongManager backend, in order to do correctly.