Fix effects chains not loading correctly from file
This commit fixes effects chains either being completely silent
or not showing any effects sounds when loaded from a file.
This commit also converts the Context in the ConnectionManager
to a GainNode, which is muted and unmuted by the SongManager
in order to prevent long delay effects from continuing after stop.